Keenan: Implemented volume slider

+++ Assets/Scripts/SettingsPanel.cs
@@ -44,7 +44,7 @@ public class SettingsPanel : MonoBehaviour
     {
         Debug.Log("Saving settings");
         PersistentData.persistentData.setSensitivity(sensitivitySlider.value);
-        PersistentData.persistentData.setVolume((int)volumeSlider.value);
+        PersistentData.persistentData.setVolume(volumeSlider.value);
         PersistentData.persistentData.savePlayerPrefs();
     }
 }
